Rewrite messages synchronization
I have intrudoced "jobIDs". When Conversation page wants to get messages,
it makes a request to messages Vuex store and gets jobID in response.
When the store is ready, it fires COMPLETE_JOB mutation with the jobID as payload.
Conversation page subscribes to this mutation and when its time, it ask
the store for results.